requirements.txt
altgraph==0.17.4
mysql-connector-python==9.1.0
packaging==24.1
pefile==2023.2.7
pillow==11.0.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyMySQL==1.1.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
